build: don't include fiber.h into a merger test
The problem is described in a previous commit. In short: inclusion of
`fiber.h` leads to inclusion of `libunwind.h`, but it is not necessarily
in include paths that used for testing sources.

We can add the include path or eliminate the `fiber.h` dependency. I
have no strict preference here. However the latter allows to write
public API and internal API calls differently in the testing code and it
looks more clean.

The implementation is based on the `tnt_internal_symbol()` function,
which was introduced to make a dark magic of this kind. See commit
395c30e8f85d ("fiber_channel: add accessor to internal functions").

Fixes #8025

NO_DOC=It is a fix of a build failure.
NO_TEST=It is a fix of a build failure.