ci: fix running full testing on any label when `full-ci` is set
Previously, setting any label on a pull request with the `full-ci` label
triggered running the full testing. The reason was in not enough strict
condition for running workflows.

Now the full testing will be run only when pull requests are labeled by
the 'full-ci' label, or it already exists, but event's activity type is
not related to labeling.

NO_DOC=ci
NO_TEST=ci
NO_CHANGELOG=ci