ci: update submodule to HEAD of remote branch
Workflow submodule_update used the HEAD of local branch on the runner
as a git reference for updating the submodule. If the repository was
already checked out on the runner before, it had the same old
`master` branch from the first checkout.
Now the workflow will force action tarantool/actions/update-submodule
to use the HEAD of remote branch. This remote branch must be up-to-date.

NO_DOC=CI
NO_TEST=CI
NO_CHANGELOG=CI