debugger: refactor debugging scenarios
Refactor the way how we run various debugger scenarios, getting rid
of external static file, and producing debuggee scripts on the fly.
The idea is to have everything in the single source: both source
script to be used for debugging, and corresponding debugger commands.

NO_DOC=internal
NO_CHANGELOG=internal