iproto: replace tx_check_schema with tx_check_msg
After accepting an iproto message in tx, we call tx_check_schema to
check if the schema version supplied by the user matches the actual
one. If it doesn't, we bail out early with ER_WRONG_SCHEMA_VERSION.

We need to add more checks that should be done before continuing with
processing a message in tx. tx_check_schema looks like a perfect place
for it. Let's rename it to tx_check_msg and pass an accepted message
instead of the schema version to it.

Needed for https://github.com/tarantool/tarantool-ee/issues/301

NO_DOC=refactoring
NO_TEST=refactoring
NO_CHANGELOG=refactoring