schema: make _schema space transactional
_schema on_replace trigger used to treat replace as commit. No
support for rollback at all and all changes are immediately
visible.

That is fine most of the time but still incorrect. The patch makes
the space properly respect transactions.

This is done as a preparation for adding several new _schema keys
which will be transactional from the start. Would be strange to
leave certain keys ignoring transactions. Hence this fix is done.

In scope of #5029

NO_DOC=bugfix
NO_CHANGELOG=couldn't happen with legal usage of public APIs