raft: fix an assertion failure in raft_stop_candidate
There is a false assertion in raft_stop_candidate(): it assumes that the
node must always have a running timer whenever it sees the leader.
This is not true when the node is busy writing the new term on disk.

Cover the mentioned case in the assertion.

Closes #8169

NO_DOC=bugfix

Co-authored-by: Sergey Ostanevich <sergos@tarantool.org>
(cherry picked from commit df6cf5ec639aa414e3cd013de3edb31f59fbbe04)