lua: don't tweak package.loaded.compat
The module can't be pulled as `require('compat')`, because it is a
tarantool specific tuning mechanism, not a general compatibility module.
The API is exposed as `require('tarantool').compat`.

Before this patch the module was loaded as `package.loaded.compat`,
captured and then removed from `package.loaded`. However, it would be
easier to follow if we'll just assign an internal name for the module.

The module is renamed to `internal.compat`. The Lua/C part of the module
is renamed to `internal.compat.lib`.

Part of #7774

NO_DOC=user visible behavior is unchanged, pure refactoring change
NO_TEST=see NO_DOC
NO_CHANGELOG=see NO_DOC