lua: implement searchers-to-loaders mapping
This patch finalizes the refactoring of Tarantool package.loaders
required for Tarantool integrity protection.

Finally, to make it possible to instrument the searchers with Tarantool
integrity protection, two steps are left.
* Implement mapping from searchers table to <package.loaders> using
  <gen_legacy_loader> helper.
* Obtain the searcher from the table directly, so the runtime
  instrumentation takes the effect.

As a result, <gen_legacy_loader> is reimplemented to use the table with
searchers as an upvalue instead of the particular searcher routine, and
<package.loaders> table is filled by the function generated with the
updated helper. Furthermore, the aforementioned table with searchers is
added to 'internal.loaders' module exports to make the instrumentation
more convenient for users.

Needed for tarantool/tarantool-ee#585

NO_DOC=refactoring
NO_TEST=refactoring
NO_CHANGELOG=refactoring