lua: implement croot searcher
This patch continues the refactoring of Tarantool package.loaders
required for Tarantool integrity protection.

Though croot loader is not the popular way to work with Lua C libraries,
some modules are organised using this mechanism, so they have to be
verified via Tarantool integrity protection system, hence, croot loader
has to be reimplemented to the new searching machinery.

The corresponding croot searcher is implemented in this patch and added
to the indexed part of the <package.searchers> table.

It's worth to mention two peculiarities related to the croot loader and
searcher:

* Since Tarantool uses the original croot loader from LuaJIT, it ignores
  all paths except ones specified in <package.cpath>. However, there
  might be some croot-like modules in the Tarantool-specific cpaths (cwd
  or application root). It looks like croot machinery is not widely
  used, if this bug bothers nobody, but it still needs to be fixed.
* A similar situation relates to the overriding mechanism that was
  recently introduced: there is no "prefixed" croot loader and all
  libraries using croot loading machinery are ignored for the overridden
  paths as a result.

Needed for tarantool/tarantool-ee#585

NO_DOC=refactoring
NO_TEST=refactoring
NO_CHANGELOG=refactoring