core: add tweaks subsystem
This commit adds an internal C API for registering tweaks. A tweak is
an object that provides a convenient setter/getter API for a global C
variable.

To register a tweak, use a TWEAK_XXX macro at the global level in
a C source file:

  static int my_var;
  TWEAK_INT(my_var);

The name of a tweak equals the name of the underlying variable
("my_var" in the example above).

To set/get a tweak value, use the tweak_set and tweak_get functions:

  struct tweak_value val;
  tweak_get("my_var", &val);
  val.ival = 42;
  tweak_set("my_var", &val);

The tweak_value struct is a variant that can contain one of three types:
int, bool, and string, one per each available tweak types:

  TWEAK_BOOL(bool_var);
  TWEAK_INT(int_var);
  TWEAK_ENUM(enum_name, enum_var);

The TWEAK_ENUM macro is special, as it also requires the name of the
enum type to be passed. When a enum tweak value is exported/imported,
it is converted to a string using the STR2ENUM macro.

It's also possible to iterate over all registered tweaks using the
tweak_foreach() function.

The tweak registry is a simple hash table mapping tweak names to
tweak objects, which in turn point to underlying variables. Tweaks
are registered at startup using the constructor function attribute.

Part of #7883
Needed for #8117

NO_DOC=internal
NO_CHANGELOG=internal