compat: switch yaml_pretty_multiline to tweaks
This commits adds the yaml_pretty_multiline variable and a tweak for it
that is now used by the compat module. The new variable configures
whether all multiline strings are encoded in the block scalar style or
only those that contain a '\n\n' substring.

Part of #8117

NO_DOC=refactoring
NO_TEST=refactoring
NO_CHANGELOG=refactoring